requirements.txt
Django==3.2.3
black==21.4b2
factory-boy==3.2.0
Faker==8.1.0
django-environ==0.7.0
psycopg2-binary==2.8.6
djangorestframework==3.13.1
